feat(esm)!: make the remaining dual-mode packages ESM-only
@cedarjs/prerender, @cedarjs/testing and @cedarjs/project-config drop their
dist/cjs builds, tsconfig.cjs.json files and `require` export conditions.
The CJS-only prerender runner and its Babel plugins are removed; the ESM runner
is now src/runPrerender.tsx. The CJS build helpers in @cedarjs/framework-tools
(buildCjs, buildExternalCjs, generateTypesCjs, insertCommonJsPackageJson) have
no consumers left and are removed, as are the legacy `apollo/` and `toast/`
CommonJS folder shims in @cedarjs/web.
