+++ packages/framework-tools/src/buildDefaults.ts
@@ -87,33 +87,12 @@ export async function build({
   }
 }
 
-export function buildCjs() {
-  return build({
-    buildOptions: {
-      ...defaultBuildOptions,
-      tsconfig: 'tsconfig.cjs.json',
-      outdir: 'dist/cjs',
-    },
-  })
-}
-
 export function buildEsm() {
   return build({
     buildOptions: {
       ...defaultBuildOptions,
       tsconfig: 'tsconfig.build.json',
       format: 'esm',
-    },
-  })
-}
-
-export function buildExternalCjs() {
-  return build({
-    buildOptions: {
-      ...defaultBuildOptions,
-      tsconfig: 'tsconfig.cjs.json',
-      outdir: 'dist/cjs',
-      packages: 'external',
     },
   })
 }
