+++ packages/prerender/build.mts
@@ -1,9 +1,5 @@
-import { build, buildCjs, defaultBuildOptions } from '@cedarjs/framework-tools'
-import {
-  generateTypesCjs,
-  generateTypesEsm,
-  insertCommonJsPackageJson,
-} from '@cedarjs/framework-tools/generateTypes'
+import { build, defaultBuildOptions } from '@cedarjs/framework-tools'
+import { generateTypesEsm } from '@cedarjs/framework-tools/generateTypes'
 
 await build({
   buildOptions: {
@@ -11,23 +7,5 @@ await build({
     tsconfig: 'tsconfig.build.json',
     format: 'esm',
   },
-  entryPointOptions: {
-    ignore: [
-      '**/__tests__',
-      '**/*.test.{ts,js}',
-      '**/__fixtures__',
-      '**/testUtils',
-      '**/__testfixtures__',
-      '**/__typetests__',
-      'src/runPrerender.tsx',
-      'src/index.ts',
-    ],
-  },
 })
 await generateTypesEsm()
-
-await buildCjs()
-await generateTypesCjs()
-await insertCommonJsPackageJson({
-  buildFileUrl: import.meta.url,
-})
