+++ packages/prerender/package.json
@@ -11,38 +11,18 @@
   "type": "module",
   "exports": {
     ".": {
-      "import": {
-        "types": "./dist/runPrerenderEsm.d.ts",
-        "default": "./dist/runPrerenderEsm.js"
-      },
-      "require": {
-        "types": "./dist/cjs/runPrerender.d.ts",
-        "default": "./dist/cjs/runPrerender.js"
-      }
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
     },
     "./browserUtils": {
-      "import": {
-        "types": "./dist/browserUtils/index.d.ts",
-        "default": "./dist/browserUtils/index.js"
-      },
-      "require": {
-        "types": "./dist/cjs/browserUtils/index.d.ts",
-        "default": "./dist/cjs/browserUtils/index.js"
-      }
+      "types": "./dist/browserUtils/index.d.ts",
+      "default": "./dist/browserUtils/index.js"
     },
     "./detection": {
-      "import": {
-        "types": "./dist/detection/detection.d.ts",
-        "default": "./dist/detection/detection.js"
-      },
-      "require": {
-        "types": "./dist/cjs/detection/detection.d.ts",
-        "default": "./dist/cjs/detection/detection.js"
-      }
+      "types": "./dist/detection/detection.d.ts",
+      "default": "./dist/detection/detection.js"
     }
   },
-  "main": "dist/cjs/index.js",
-  "types": "./dist/cjs/index.d.ts",
   "files": [
     "dist"
   ],
@@ -50,7 +30,6 @@
     "build": "node ./build.mts && yarn build:types",
     "build:pack": "yarn pack -o cedarjs-prerender.tgz",
     "build:types": "tsc --build --verbose ./tsconfig.build.json",
-    "build:types-cjs": "tsc --build --verbose ./tsconfig.cjs.json",
     "build:watch": "nodemon --watch src --ext \"js,jsx,ts,tsx,template\" --ignore dist --exec \"yarn build\"",
     "check:attw": "yarn cedar-fwtools-attw",
     "check:package": "concurrently npm:check:attw yarn:publint",
@@ -88,7 +67,6 @@
     "@cedarjs/web": "workspace:*",
     "@types/mime-types": "2.1.4",
     "@types/react": "^19.2.0",
-    "babel-plugin-tester": "11.0.4",
     "concurrently": "9.2.4",
     "publint": "0.3.23",
     "typescript": "5.9.3",
