--- packages/prerender/src/babelPlugins/__tests__/__fixtures__/cell/cell-with-default-and-other-named-export/code.js
@@ -1,7 +1,0 @@
-export const Success = () => {
-  return 'string'
-}
-
-export default () => {
-  return 'string'
-}
