--- packages/prerender/src/babelPlugins/__tests__/__fixtures__/cell/cell-with-default-export/code.js
@@ -1,3 +1,0 @@
-export default () => {
-  return 'string'
-}
