--- packages/prerender/src/babelPlugins/__tests__/__fixtures__/cell/fragment-cell/code.js
@@ -1,10 +1,0 @@
-export const FRAGMENT = gql`
-  fragment FragmentCell_author on User {
-    email
-    fullName
-  }
-`
-
-export const Success = ({ author }) => {
-  return JSON.stringify(author, null, 2)
-}
