--- packages/prerender/src/babelPlugins/__tests__/babel-plugin-redwood-cell.test.ts
@@ -1,11 +1,0 @@
-import path from 'path'
-
-import pluginTester from 'babel-plugin-tester'
-
-import { babelPluginRedwoodCell } from '../babel-plugin-redwood-cell'
-
-pluginTester({
-  plugin: babelPluginRedwoodCell,
-  pluginName: 'babel-plugin-redwood-cell',
-  fixtures: path.join(__dirname, '__fixtures__/cell'),
-})
