+++ packages/prerender/src/runPrerender.tsx
@@ -2,7 +2,7 @@ import fs from 'node:fs'
 import path from 'node:path'
 
 import React from 'react'
-import type { ElementType } from 'react'
+import type { ElementType, FunctionComponent } from 'react'
 
 import { ApolloClient, ApolloLink, InMemoryCache } from '@apollo/client'
 import type { CheerioAPI } from 'cheerio'
@@ -10,15 +10,15 @@ import { load as loadHtml } from 'cheerio'
 import ReactDOMServer from 'react-dom/server'
 
 import {
-  registerApiSideBabelHook,
-  registerWebSideBabelHook,
-} from '@cedarjs/babel-config'
-import { getPaths, ensurePosixPath } from '@cedarjs/project-config'
+  getPaths,
+  ensurePosixPath,
+  importStatementPath,
+} from '@cedarjs/project-config'
+import { matchPath } from '@cedarjs/router/dist/util'
 import type { QueryInfo } from '@cedarjs/web'
 
 import { hasApolloState } from './apolloState.js'
-import { babelPluginRedwoodCell } from './babelPlugins/babel-plugin-redwood-cell.js'
-import { babelPluginRedwoodPrerenderMediaImports } from './babelPlugins/babel-plugin-redwood-prerender-media-imports.js'
+import { buildAndImport } from './build-and-import/buildAndImport.js'
 import { detectPrerenderRoutes } from './detection/detection.js'
 import {
   GqlHandlerImportError,
@@ -27,6 +27,7 @@ import {
 } from './errors.js'
 import { executeQuery, getGqlHandler } from './graphql/graphql.js'
 import type { FileImporter } from './graphql/graphql.js'
+import { NodeRunner } from './graphql/node-runner.js'
 import { getRootHtmlPath, registerShims, writeToDist } from './internal.js'
 
 // Create an apollo client that we can use to prepopulate the cache and restore it client-side
@@ -40,14 +41,15 @@ const prerenderApolloClient = new ApolloClient({
 async function recursivelyRender(
   App: ElementType,
   Routes: ElementType,
+  CellCacheContextProvider: ElementType,
+  LocationProvider: ElementType,
   renderPath: string,
   fileImporter: FileImporter,
   gqlHandler: any,
   queryCache: Record<string, QueryInfo>,
 ): Promise<string> {
   // Load this async, to prevent rwjs/web being loaded before shims
-  const { CellCacheContextProvider, getOperationName } = require('@cedarjs/web')
-  const { LocationProvider } = require('@cedarjs/router')
+  const { getOperationName } = await import('@cedarjs/web')
 
   let shouldShowGraphqlHandlerNotFoundWarn = false
   // Execute all gql queries we haven't already fetched
@@ -148,6 +150,8 @@ async function recursivelyRender(
     return recursivelyRender(
       App,
       Routes,
+      CellCacheContextProvider,
+      LocationProvider,
       renderPath,
       fileImporter,
       gqlHandler,
@@ -168,9 +172,6 @@ function insertChunkLoadingScript(
   indexHtmlTree: CheerioAPI,
   renderPath: string,
 ) {
-  // Load this async, to prevent rwjs/router being loaded before shims
-  const { matchPath } = require('@cedarjs/router/dist/util')
-
   const prerenderRoutes = detectPrerenderRoutes()
 
   const route = prerenderRoutes.find((route) => {
@@ -199,9 +200,6 @@ function insertChunkLoadingScript(
   const chunkPaths: string[] = []
 
   if (route?.filePath) {
-    // + 4 skips the 'web/' prefix (4 chars). The resulting pagePath
-    // includes 'src/pages/...' which matches the build manifest keys which are
-    // relative to Vite's root, which is web/
     const pagesIndex =
       route.filePath.indexOf(path.join('web', 'src', 'pages')) + 4
     const pagePath = ensurePosixPath(route.filePath.slice(pagesIndex))
@@ -251,6 +249,35 @@ function insertChunkLoadingScript(
   })
 }
 
+interface Args {
+  appPath: string
+  routesPath: string
+  outDir: string
+}
+
+async function createCombinedEntry({ appPath, routesPath, outDir }: Args) {
+  const combinedContent = `
+    import { LocationProvider } from '@cedarjs/router'
+    import { CellCacheContextProvider } from '@cedarjs/web'
+
+    import App from "${importStatementPath(appPath.replace('.tsx', ''))}";
+    import Routes from "${importStatementPath(routesPath.replace('.tsx', ''))}";
+
+    export { LocationProvider, CellCacheContextProvider, App, Routes };
+  `
+
+  const tempFilePath = path.join(outDir, '__prerender-temp-entry.tsx')
+  await fs.promises.writeFile(tempFilePath, combinedContent, 'utf8')
+  return tempFilePath
+}
+
+const renderCache: {
+  App?: FunctionComponent
+  Routes?: FunctionComponent
+  CellCacheContextProvider?: FunctionComponent
+  LocationProvider?: FunctionComponent
+} = {}
+
 interface PrerenderParams {
   queryCache: Record<string, QueryInfo>
   renderPath: string // The path (url) to render e.g. /about, /dashboard/me, /blog-post/3
@@ -262,80 +289,68 @@ export const runPrerender = async ({
 }: PrerenderParams): Promise<string | void> => {
   registerShims(renderPath)
 
-  // registerApiSideBabelHook already includes the default api side babel
-  // config. So what we define here is additions to the default config
-  registerApiSideBabelHook({
-    plugins: [
-      [
-        'babel-plugin-module-resolver',
-        {
-          alias: {
-            api: getPaths().api.base,
-            web: getPaths().web.base,
-          },
-          loglevel: 'silent', // to silence the unnecessary warnings
-        },
-      ],
-    ],
-    overrides: [
-      {
-        test: ['./api/'],
-        plugins: [
-          [
-            'babel-plugin-module-resolver',
-            {
-              alias: {
-                src: getPaths().api.src,
-              },
-              loglevel: 'silent',
-            },
-            'exec-api-src-module-resolver',
-          ],
-        ],
-      },
-    ],
-  })
+  const nodeRunner = new NodeRunner()
 
-  const gqlHandler = await getGqlHandler(require)
+  const gqlHandler = await getGqlHandler(nodeRunner.importFile.bind(nodeRunner))
 
-  // Prerender specific configuration
-  // extends projects web/babelConfig
-  registerWebSideBabelHook({
-    overrides: [
-      {
-        plugins: [
-          // Vite handles CSS imports for the browser build
-          ['ignore-html-and-css-imports'],
-          [babelPluginRedwoodPrerenderMediaImports],
-        ],
-      },
-      {
-        test: /.+Cell.(js|tsx|jsx)$/,
-        plugins: [babelPluginRedwoodCell],
-      },
-    ],
-    options: {
-      forPrerender: true,
-    },
-  })
+  const prerenderDistPath = path.join(getPaths().web.dist, '__prerender')
+  fs.mkdirSync(prerenderDistPath, { recursive: true })
 
-  const indexContent = fs.readFileSync(getRootHtmlPath()).toString()
-  const { default: App } = require(getPaths().web.app)
-  const { default: Routes } = require(getPaths().web.routes)
+  if (!renderCache.App) {
+    const entryPath = await createCombinedEntry({
+      appPath: getPaths().web.app,
+      routesPath: getPaths().web.routes,
+      outDir: prerenderDistPath,
+    })
+
+    const required = await buildAndImport({
+      filepath: entryPath,
+      preserveTemporaryFile: true,
+    })
+
+    renderCache.App = required.App
+    renderCache.Routes = required.Routes
+    renderCache.CellCacheContextProvider = required.CellCacheContextProvider
+    renderCache.LocationProvider = required.LocationProvider
+  }
+
+  const { LocationProvider, App, Routes, CellCacheContextProvider } =
+    renderCache
+
+  if (!App) {
+    throw new Error('App not found')
+  }
+
+  if (!Routes) {
+    throw new Error('Routes not found')
+  }
+
+  if (!CellCacheContextProvider) {
+    throw new Error('CellCacheContextProvider not found')
+  }
+
+  if (!LocationProvider) {
+    throw new Error('LocationProvider not found')
+  }
 
   const componentAsHtml = await recursivelyRender(
     App,
     Routes,
+    CellCacheContextProvider,
+    LocationProvider,
     renderPath,
-    require,
+    nodeRunner.importFile.bind(nodeRunner),
     gqlHandler,
     queryCache,
   )
 
+  nodeRunner.close()
+
   const { helmet } = globalThis.__REDWOOD__HELMET_CONTEXT
 
-  // Loop over ther queryCache and write the queries to the apollo client cache this will normalize the data
-  // and make it available to the app when it hydrates
+  // Loop over ther queryCache and write the queries to the apollo client cache
+  // This will normalize the data and make it available to the app when it
+  // hydrates
   Object.keys(queryCache).forEach((queryKey) => {
     const { query, variables, data } = queryCache[queryKey]
     prerenderApolloClient.writeQuery({
@@ -345,6 +360,7 @@ export const runPrerender = async ({
     })
   })
 
+  const indexContent = fs.readFileSync(getRootHtmlPath()).toString()
   const indexHtmlTree = loadHtml(indexContent)
 
   if (helmet) {
