--- packages/prerender/tsconfig.cjs.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "outDir": "dist/cjs",
-    "tsBuildInfoFile": "./tsconfig.cjs.tsbuildinfo",
-    "verbatimModuleSyntax": false
-  },
-  "exclude": [
-    "**/__tests__",
-    "src/runPrerenderEsm.tsx",
-    "src/build-and-import",
-    "src/graphql/node-runner.ts",
-    "src/graphql/vite*"
-  ]
-}
