+++ packages/project-config/package.json
@@ -10,34 +10,16 @@
   "type": "module",
   "exports": {
     ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/cjs/index.d.ts",
-        "default": "./dist/cjs/index.js"
-      }
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
     },
     "./packageManager": {
-      "import": {
-        "types": "./dist/packageManager.d.ts",
-        "default": "./dist/packageManager.js"
-      },
-      "default": {
-        "types": "./dist/cjs/packageManager.d.ts",
-        "default": "./dist/cjs/packageManager.js"
-      }
+      "types": "./dist/packageManager.d.ts",
+      "default": "./dist/packageManager.js"
     },
     "./workspaces": {
-      "import": {
-        "types": "./dist/workspaces.d.ts",
-        "default": "./dist/workspaces.js"
-      },
-      "default": {
-        "types": "./dist/cjs/workspaces.d.ts",
-        "default": "./dist/cjs/workspaces.js"
-      }
+      "types": "./dist/workspaces.d.ts",
+      "default": "./dist/workspaces.js"
     }
   },
   "types": "./dist/index.d.ts",
@@ -47,7 +29,7 @@
   "scripts": {
     "build": "node ./build.ts && run build:types",
     "build:pack": "yarn pack -o cedarjs-project-config.tgz",
-    "build:types": "tsc --build --verbose ./tsconfig.json ./tsconfig.cjs.json",
+    "build:types": "tsc --build --verbose ./tsconfig.json",
     "build:watch": "nodemon --watch src --ext \"js,ts,tsx\" --ignore dist --exec \"yarn build\"",
     "check:attw": "yarn cedar-fwtools-attw",
     "check:package": "concurrently npm:check:attw yarn:publint",
