--- packages/testing/tsconfig.cjs.json
@@ -1,17 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "outDir": "dist/cjs",
-    "tsBuildInfoFile": "./tsconfig.cjs.tsbuildinfo",
-    "verbatimModuleSyntax": false
-  },
-  "exclude": [
-    "dist",
-    "node_modules",
-    "**/__tests__",
-    "**/__mocks__",
-    "**/*.test.*",
-    "src/api/vitest",
-    "src/web/vitest"
-  ]
-}
