--- packages/web/apollo/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/cjs/apollo/index.js')
