--- packages/web/toast/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/toast')
