+++ packages/web/vitest.config.mts
@@ -5,9 +5,8 @@ export default defineConfig({
     exclude: [
       ...configDefaults.exclude,
       // Vitest 4 removed `**/dist/**` from its default excludes. Compiled
-      // copies of the .tsx test files end up in dist (and dist/cjs, where
-      // they'd crash since Vitest can't be require()d from CJS), so keep
-      // excluding them like Vitest 3 did
+      // copies of the .tsx test files end up in dist, so keep excluding them
+      // like Vitest 3 did
       '**/dist/**',
       '**/fixtures',
       '**/__typetests__',
