+++ tasks/framework-tools/tarsync/lib.mts
@@ -37,7 +37,6 @@ export const IGNORED = [
 
   /tsconfig.tsbuildinfo/,
   /tsconfig.build.tsbuildinfo/,
-  /tsconfig.cjs.tsbuildinfo/,
 
   // The tarballs generated by `yarn build:pack`
   /cedar-.*\.tgz$/,
